requirements.txt
flask
google-generativeai
Pillow
google-cloud-speech